refactor: moving mssql into the new client
