+++ azurerm/internal/provider/services.go
@@ -37,6 +37,7 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/media"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/monitor"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/msi"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/mssql"
 )
 
 func SupportedServices() []common.ServiceRegistration {
@@ -76,5 +77,6 @@ func SupportedServices() []common.ServiceRegistration {
 		media.Registration{},
 		monitor.Registration{},
 		msi.Registration{},
+		mssql.Registration{},
 	}
 }
