+++ azurerm/internal/services/mssql/client/client.go
@@ -1,4 +1,4 @@
-package mssql
+package client
 
 import (
 	"github.com/Azure/azure-sdk-for-go/services/preview/sql/mgmt/2017-10-01-preview/sql"
@@ -9,7 +9,7 @@ type Client struct {
 	ElasticPoolsClient *sql.ElasticPoolsClient
 }
 
-func BuildClient(o *common.ClientOptions) *Client {
+func NewClient(o *common.ClientOptions) *Client {
 	ElasticPoolsClient := sql.NewElasticPoolsClientWithBaseURI(o.ResourceManagerEndpoint, o.SubscriptionId)
 	o.ConfigureClient(&ElasticPoolsClient.Client, o.ResourceManagerAuthorizer)
